requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
audioplayer==0.6
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
distro==1.9.0
edge-tts==6.1.12
flatbuffers==24.3.25
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
joblib==1.4.2
mpmath==1.3.0
multidict==6.0.5
numpy==1.26.4
onnxruntime==1.18.0
openai==1.30.3
openwakeword==0.4.0
packaging==24.0
protobuf==5.27.0
PyAudio==0.2.14
pycairo==1.26.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
PyGObject==3.48.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
requests==2.32.2
scikit-learn==1.5.0
scipy==1.13.1
simple-websocket==1.0.0
sniffio==1.3.1
SpeechRecognition==3.10.4
sympy==1.12
threadpoolctl==3.5.0
tqdm==4.66.4
translate-shell==0.0.51
typing_extensions==4.12.0
urllib3==2.2.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
